backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==21.2.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
fairlearn==0.10.0
shap==0.45.1
matplotlib==3.9.0
firebase-admin==6.5.0
google-generativeai==0.7.2
python-dotenv==1.0.1
Werkzeug==3.0.3
firebase-functions==0.1.1
